Изменил типы

+++ src/main/java/edu/project3/reporters/LogReporter.java
@@ -48,7 +48,7 @@ public class LogReporter {
             resource,
             from.format(formatter),
             to.format(formatter),
-            logs.length,
+            (long) logs.length,
             calculateAverageResponseSize()
         );
     }
